inp_str, out_str: totally rewritten according to the format of strtoc;    inp_str sets the number to nan+i*nan in case of an error
git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/mpc/trunk@483 211d60ee-9f03-0410-a15a-8952a2c7a4e4

+++ src/get_str.c
@@ -42,7 +42,7 @@ MA 02111-1307, USA. */
 #endif
 
 static char *
-mpc_alloc_str (int len) {
+mpc_alloc_str (size_t len) {
    void * (*allocfunc) (size_t);
    mp_get_memory_functions (&allocfunc, NULL, NULL);
    return (char *) ((*allocfunc) (len * sizeof (char)));
